Updated EverApp ts file to store EverApp related information
