+++ assets/src/apps/delivery/layouts/deck/DeckLayoutHeader.tsx
@@ -3,11 +3,10 @@ import React, { useState } from 'react';
 import { useSelector } from 'react-redux';
 import { selectScore } from '../../store/features/page/slice';
 import { OverlayTrigger, Popover } from 'react-bootstrap';
-import { BeagleAppIFrame } from '../deck/BeagleAppIFrame';
+import { EverAppActivity, getEverAppActivity, udpateAttemptGuid } from './EverApps';
 import { selectPageContent } from '../../store/features/page/slice';
-import ActivityRenderer from 'apps/delivery/components/ActivityRenderer';
 import { ActivityState } from 'components/activities/types';
-import { clone } from 'utils/common';
+import ActivityRenderer from 'apps/delivery/components/ActivityRenderer';
 import BeagleLogo from '../../../../../static/images/icons/icon-nine-dots.svg';
 import { defaultGlobalEnv, evalScript } from '../../../../adaptivity/scripting';
 
@@ -35,44 +34,44 @@ const DeckLayoutHeader: React.FC<DeckLayoutHeaderProps> = ({
   const currentPage = useSelector(selectPageContent);
   const everApps = currentPage?.custom?.everApps;
 
-  const activityState: ActivityState = {
-    attemptGuid: 'preview_2946819616',
-    attemptNumber: 1,
-    dateEvaluated: null,
-    score: null,
-    outOf: null,
-    parts: [
-      {
-        attemptGuid: 'sampleIframeGuid',
-        attemptNumber: 1,
-        dateEvaluated: null,
-        score: null,
-        outOf: null,
-        response: null,
-        feedback: null,
-        hints: [],
-        partId: 'janus_capi_iframe-3311152192',
-        hasMoreAttempts: false,
-        hasMoreHints: false,
-      },
-    ],
-    hasMoreAttempts: true,
-    hasMoreHints: true,
-  };
+  // const activityState: ActivityState = {
+  //   attemptGuid: 'preview_2946819616',
+  //   attemptNumber: 1,
+  //   dateEvaluated: null,
+  //   score: null,
+  //   outOf: null,
+  //   parts: [
+  //     {
+  //       attemptGuid: 'sampleIframeGuid',
+  //       attemptNumber: 1,
+  //       dateEvaluated: null,
+  //       score: null,
+  //       outOf: null,
+  //       response: null,
+  //       feedback: null,
+  //       hints: [],
+  //       partId: 'janus_capi_iframe-3311152192',
+  //       hasMoreAttempts: false,
+  //       hasMoreHints: false,
+  //     },
+  //   ],
+  //   hasMoreAttempts: true,
+  //   hasMoreHints: true,
+  // };
 
-  const updateEverAppIFrameURL = (everAppObj: any, url: string, index: number) => {
-    const updatedObject = clone(everAppObj);
-    updatedObject.id = everAppObj.id + index;
-    updatedObject.attemptGuid = everAppObj.attemptGuid + index;
-    updatedObject.content.partsLayout[0].custom.src = url;
-    return updatedObject;
-  };
+  // const updateEverAppIFrameURL = (everAppObj: any, url: string, index: number) => {
+  //   const updatedObject = clone(everAppObj);
+  //   updatedObject.id = everAppObj.id + index;
+  //   updatedObject.attemptGuid = everAppObj.attemptGuid + index;
+  //   updatedObject.content.partsLayout[0].custom.src = url;
+  //   return updatedObject;
+  // };
 
-  const udpateAttemptGuid = (index: number) => {
-    const updatedObject = clone(activityState);
-    updatedObject.attemptGuid = activityState.attemptGuid + index;
-    return updatedObject;
-  };
+  // const udpateAttemptGuid = (index: number) => {
+  //   const updatedObject = clone(everAppActivityState);
+  //   updatedObject.attemptGuid = everAppActivityState.attemptGuid + index;
+  //   return updatedObject;
+  // };
 
   return (
     <div className="headerContainer">
@@ -122,8 +121,8 @@ const DeckLayoutHeader: React.FC<DeckLayoutHeaderProps> = ({
                                     <div style={{ zIndex: 9999, opacity: 1, marginTop: '5px' }}>
                                       <ActivityRenderer
                                         key={everApp.id}
-                                        activity={updateEverAppIFrameURL(
-                                          BeagleAppIFrame,
+                                        activity={getEverAppActivity(
+                                          EverAppActivity,
                                           everApp.url,
                                           index,
                                         )}
